refactor(postpage): ファイルアップロード側のコンポーネント群を適切な大きさで分類する
