+++ app/components/post/elements/after-upload-drop-zone-content.tsx
@@ -1,0 +1,26 @@
+import { LiaCheckCircle } from "react-icons/lia";
+import { EnvFileSelectButton } from "./env-file-select-button";
+
+interface AfterUploadDropZoneContentProps {
+  open: () => void;
+  dropFiles: File[];
+}
+
+export function AfterUploadDropZoneContent({
+  open,
+  dropFiles,
+}: AfterUploadDropZoneContentProps) {
+  return (
+    <>
+      <LiaCheckCircle size={40} color="skyblue" />
+      {dropFiles.length >= 2 ? (
+        <span>
+          {dropFiles[0].name} ...他{dropFiles.length - 1}件
+        </span>
+      ) : (
+        <span>{dropFiles[0].name}</span>
+      )}
+      <EnvFileSelectButton open={open} />
+    </>
+  );
+}
