+++ app/hooks/post/usePersonalOrOrganization.ts
@@ -1,0 +1,9 @@
+import { useState } from "react";
+
+export function usePersonalOrOrganization() {
+  const [isPersonalClicked, setIsPersonalClicked] = useState<boolean>(true);
+  const clickPersonal = () => setIsPersonalClicked(true);
+  const clickOrganization = () => setIsPersonalClicked(false);
+
+  return { isPersonalClicked, clickPersonal, clickOrganization };
+}
