+++ app/repository-view/page.tsx
@@ -1,9 +1,5 @@
 "use client";
-import { useEffect, useState } from "react";
-import { IconType } from "react-icons";
-import { FaUser } from "react-icons/fa";
-import { FaUsers } from "react-icons/fa";
-import { TbZoom } from "react-icons/tb";
+import { Dispatch, SetStateAction, useEffect, useState } from "react";
 import { FcFolder } from "react-icons/fc";
 import { IoLogoGithub, IoMdClose } from "react-icons/io";
 
@@ -32,6 +28,12 @@ import { AiOutlineClose } from "react-icons/ai";
 
 import Modal from "react-modal";
 import { useModal } from "../hooks/toppage/useModal";
+import { usePersonalOrOrganization } from "../hooks/post/usePersonalOrOrganization";
+import { RepoNameSearchBar } from "../components/post/elements/repo-name-search-bar";
+import { PersonalOrganizationSelctor } from "../components/post/elements/personal-organization-selector";
+import { RepositorySelctorArea } from "../components/post/layout/repository-selector-area";
+import { DropEnvArea } from "../components/post/layout/drop-env-area";
+import { UploadPreviewFileContent } from "../components/post/elements/upload-preview-file-content";
 
 const modalStyle = {
   overlay: {
@@ -64,10 +66,6 @@ const orgsData = [
 
 // TODO:コンポーネント分割・カスタムフックスの作成
 export default function Page() {
-  const [isPersonalClicked, setIsPersonalClicked] = useState<boolean>(true);
-  const clickPersonal = () => setIsPersonalClicked(true);
-  const clickOrganizations = () => setIsPersonalClicked(false);
-
   const [dropFiles, setDropFiles] = useState<File[]>([]);
   const onDrop = (dropFiles: File[]) => {
     console.log(dropFiles);
@@ -82,349 +80,11 @@ export default function Page() {
   return (
     <div className="flex">
       <div className="w-1/4 bg-white h-screen border-r border-black">
-        <RepositoryTypeSelctor
-          isPersonalClicked={isPersonalClicked}
-          clickPersonal={clickPersonal}
-          clickOrganizations={clickOrganizations}
-        />
-        <RepoNameSearchBar />
-        {repoNamesData.map((repoName, index) => (
-          <RepoSelectButton key={index} repoName={repoName} />
-        ))}
-        <OrganizationsAccordionWrapper />
+        <RepositorySelctorArea />
       </div>
       <div className="w-3/4 bg-white">
-        <div
-          className="border-2 border-dashed hover:border-gray-400 hover:bg-gray-200 transition-colors duration-500 rounded-xl mx-auto w-11/12 h-80 flex justify-center items-center mt-20"
-          {...getRootProps()}
-        >
-          <input {...getInputProps()} />
-          <div className="flex flex-col items-center">
-            {dropFiles.length === 0 ? (
-              <>
-                <FaFileUpload size={40} />
-                <span className="mt-4">
-                  ここにenvファイルをドラッグ & ドロップしてください
-                </span>
-                <span className="text-gray-400 text-sm">または</span>
-                <button
-                  className="bg-black text-white rounded-lg py-2 px-4 mt-2"
-                  onClick={open}
-                >
-                  envファイルを選択
-                </button>
-              </>
-            ) : (
-              <>
-                <LiaCheckCircle size={40} color="skyblue" />
-                {dropFiles.length >= 2 ? (
-                  <span>
-                    {dropFiles[0].name} ...他{dropFiles.length - 1}件
-                  </span>
-                ) : (
-                  <span>{dropFiles[0].name}</span>
-                )}
-                <EnvFileSelectButton open={open} />
-              </>
-            )}
-          </div>
-        </div>
-        <div className="w-11/12 mx-auto mt-10 border-2 rounded-lg p-4">
-          <p className="font-bold mb-2">アップロード中のファイル:</p>
-          {dropFiles.length === 0 ? (
-            <p>アップロード中のファイルはありません。</p>
-          ) : (
-            <>
-              <ProgressBar />
-              {dropFiles.map((dropFile, index) => (
-                <div key={index} className="border flex p-3 justify-between">
-                  <div className="flex items-center">
-                    <FaRegFile size={20} className="mr-3" />
-                    <span className="text-sm">{dropFile.name}</span>
-                  </div>
-                  <AiOutlineClose
-                    size={20}
-                    className="hover:bg-gray-300 rounded-full transition-colors duration-500"
-                    onClick={() =>
-                      setDropFiles((prev) => prev.splice(index, 1))
-                    }
-                  />
-                </div>
-              ))}
-              {/* <button className="bg-gray-800 text-white rounded-md w-full py-3 text-sm hover:bg-gray-700 transition-colors duration-300">
-                アップロードを確定する
-              </button> */}
-              <UploadConfirmButton uploadFiles={dropFiles} />
-            </>
-          )}
-        </div>
+        <DropEnvArea />
       </div>
     </div>
-  );
-}
-
-interface RepositoryTypeSelctorProps {
-  isPersonalClicked: boolean;
-  clickPersonal: () => void;
-  clickOrganizations: () => void;
-}
-
-export function RepositoryTypeSelctor({
-  isPersonalClicked,
-  clickPersonal,
-  clickOrganizations,
-}: RepositoryTypeSelctorProps) {
-  return (
-    <div className="w-11/12 bg-gray-100 mx-auto p-1 rounded-sm flex">
-      <PersonalRepositoryTypeButton
-        isPersonalClicked={isPersonalClicked}
-        clickPersonal={clickPersonal}
-        DrawingIcon={FaUser}
-        repositoryTypeText="個人"
-      />
-      <OrganizationsRepositoryTypeButton
-        isPersonalClicked={isPersonalClicked}
-        clickOrganizations={clickOrganizations}
-        DrawingIcon={FaUsers}
-        repositoryTypeText="組織"
-      />
-    </div>
-  );
-}
-
-// TODO:共通化する
-interface RepositoryTypeButtonProps {
-  isPersonalClicked: boolean;
-  clickPersonal: () => void;
-  clickOrganizations: () => void;
-  DrawingIcon: IconType;
-  repositoryTypeText: string;
-}
-
-type PersonalRepositoryTypeButtonProps = Omit<
-  RepositoryTypeButtonProps,
-  "clickOrganizations"
->;
-
-export function PersonalRepositoryTypeButton({
-  isPersonalClicked,
-  clickPersonal,
-  DrawingIcon,
-  repositoryTypeText,
-}: PersonalRepositoryTypeButtonProps) {
-  return (
-    <div
-      className={`w-1/2 py-2 flex items-center justify-center rounded ${
-        isPersonalClicked && "bg-white"
-      }`}
-      onClick={clickPersonal}
-    >
-      <DrawingIcon className="mr-2" />
-      <span className="text-sm text-gray-500">{repositoryTypeText}</span>
-    </div>
-  );
-}
-
-type OrganizationsRepositoryTypeButtonProps = Omit<
-  RepositoryTypeButtonProps,
-  "clickPersonal"
->;
-
-export function OrganizationsRepositoryTypeButton({
-  isPersonalClicked,
-  clickOrganizations,
-  DrawingIcon,
-  repositoryTypeText,
-}: OrganizationsRepositoryTypeButtonProps) {
-  return (
-    <div
-      className={`w-1/2 py-2 flex items-center justify-center rounded  ${
-        !isPersonalClicked && "bg-white"
-      }`}
-      onClick={clickOrganizations}
-    >
-      <DrawingIcon className="mr-2" />
-      <span className="text-sm text-gray-500">{repositoryTypeText}</span>
-    </div>
-  );
-}
-
-export function RepoNameSearchBar() {
-  return (
-    <div className="relative w-11/12 mx-auto">
-      <TbZoom className="absolute left-3 top-[11px]" size={20} />
-      <input
-        className="w-full flex items-center border py-2 px-10 rounded"
-        placeholder="リポジトリ名で検索"
-      />
-    </div>
-  );
-}
-
-interface RepoSelectButtonProps {
-  repoName: string;
-}
-
-export function RepoSelectButton({ repoName }: RepoSelectButtonProps) {
-  return (
-    <div className="w-11/12 mx-auto border rounded p-2 flex items-center text-sm">
-      <FcFolder className="mr-2" size={20} />
-      <span>{repoName}</span>
-    </div>
-  );
-}
-
-interface OrganizationsSelectButtonProps {
-  organizationName: string;
-  isAccordionClicked: boolean;
-}
-
-export function OrganizationsSelectButton({
-  organizationName,
-  isAccordionClicked,
-}: OrganizationsSelectButtonProps) {
-  return (
-    <div className="w-11/12 mx-auto border rounded flex items-center p-2 justify-between">
-      <IoLogoGithub className="mr-2" size={20} />
-      <span>{organizationName}</span>
-      {isAccordionClicked ? (
-        <MdKeyboardDoubleArrowDown size={25} />
-      ) : (
-        <MdKeyboardDoubleArrowRight size={25} />
-      )}
-    </div>
-  );
-}
-
-export function OrganizationsAccordionWrapper() {
-  const [toggleOrgName, setToggleOrgName] = useState<string>("");
-
-  const toggleOrgNameHandler = (currentClickOrgName: string) => {
-    if (toggleOrgName === currentClickOrgName) {
-      setToggleOrgName("");
-      return;
-    }
-
-    setToggleOrgName(currentClickOrgName);
-  };
-
-  return (
-    <Accordion allowZeroExpanded>
-      {orgsData.map((orgData, index) => (
-        <AccordionItem key={index}>
-          <AccordionItemHeading
-            onClick={() => toggleOrgNameHandler(orgData.orgName)}
-          >
-            <AccordionItemButton>
-              <OrganizationsSelectButton
-                organizationName={orgData.orgName}
-                isAccordionClicked={orgData.orgName === toggleOrgName}
-              />
-            </AccordionItemButton>
-          </AccordionItemHeading>
-          <AccordionItemPanel>
-            {orgData.repoName.map((repoName, index) => (
-              <RepoSelectButton key={index} repoName={repoName} />
-            ))}
-          </AccordionItemPanel>
-        </AccordionItem>
-      ))}
-      {/* </AccordionItem> */}
-    </Accordion>
-  );
-}
-
-export function EnvFileSelectButton({ open }: { open: () => void }) {
-  return (
-    <button
-      className="bg-black text-white rounded-lg py-2 px-4 mt-2"
-      onClick={open}
-    >
-      envファイルを選択
-    </button>
-  );
-}
-
-export function ProgressBar() {
-  const [uploadFileLength, setUploadFileLength] = useState<number>(5);
-  const [progressPercent, setProgressPercent] = useState<number>(0);
-
-  useEffect(() => {
-    const timer = setInterval(() => {
-      if (progressPercent >= 100) {
-        clearInterval(timer);
-        return;
-      }
-
-      // TODO:アップロードされたファイル数に比例してprogressPercentの値を増やす
-      setProgressPercent((prev) => prev + 20);
-    }, 1000);
-
-    return () => {
-      clearInterval(timer);
-    };
-  }, [progressPercent, uploadFileLength]);
-
-  return <Progress percent={progressPercent} />;
-}
-
-export function UploadConfirmButton({ uploadFiles }: { uploadFiles: File[] }) {
-  const { isModalOpen, modalOpen, modalClose } = useModal();
-
-  // TODO:後でファイル分割する
-  const fileSizeByteToKB = (fileSizeByte: number): string => {
-    if (fileSizeByte < 1024) {
-      return `${fileSizeByte} Byte`;
-    }
-    const fileSizeInKB = (fileSizeByte / 1024).toFixed(2);
-    return `${fileSizeInKB} KB`;
-  };
-
-  return (
-    <>
-      <button
-        className="bg-gray-800 text-white rounded-md w-full py-3 text-sm hover:bg-gray-700 transition-colors duration-300"
-        onClick={modalOpen}
-      >
-        アップロードを確定する
-      </button>
-      <Modal
-        isOpen={isModalOpen}
-        style={modalStyle}
-        onRequestClose={modalClose}
-      >
-        <div className="flex justify-between">
-          <span className="font-bold text-lg">アップロードの確認</span>
-          <IoMdClose
-            size={35}
-            onClick={modalClose}
-            className="hover:bg-gray-200 rounded-full transition-colors duration-200 p-1.5 relative -top-2 left-2"
-          />
-        </div>
-        <p className="text-sm">
-          以下のファイルをアップロードします。確認してください。
-        </p>
-        <div className="mt-4">
-          {uploadFiles.map((uploadFile, index) => (
-            <div key={index} className="flex justify-between">
-              <span>{uploadFile.name}</span>
-              <span>{fileSizeByteToKB(uploadFile.size)}</span>
-            </div>
-          ))}
-        </div>
-        <hr className="mt-2 h-2" />
-        <div className="flex justify-end space-x-2">
-          <button
-            onClick={modalClose}
-            className="border px-4 py-2 rounded-lg text-sm hover:bg-gray-200 transition-colors duration-300"
-          >
-            キャンセル
-          </button>
-          <button className="bg-black text-white px-4 py-2 rounded-lg text-sm hover:bg-gray-700 transition-colors duration-300">
-            アップロードを確定
-          </button>
-        </div>
-      </Modal>
-    </>
   );
 }
