+++ types/repo-type-button-type.ts
@@ -1,0 +1,9 @@
+import { IconType } from "react-icons";
+
+export interface RepositoryTypeButtonProps {
+  isPersonalClicked: boolean;
+  clickPersonal: () => void;
+  clickOrganizations: () => void;
+  DrawingIcon: IconType;
+  repositoryTypeText: string;
+}
